Consolidating t84 bootloaders
I have a feeling that the builds of micronucleus that are in this
directory were built for a VID/PID that Shawn changed in the bootloader,
so I'm going to just use his bootloader for the Atto84. To avoid
confusion, I renamed the bootloader file "t84_sparkfun" instead of
"t84_roshamglo" and I believe I changed all references to it.