Splitting Tiny84 Bootloaders again
Okay, that didn't work. Well, it did, but Shawn's bootloader config for
the roshamglo uses a jumper to go into bootloader mode which I don't
have on the Atto84. I rebuilt the t84_default bootloader config with the
VID/PID that is set in the included build of micronucleus.exe so
hopefully everything will work as intended now.