Add "x86_64-pc-linux-gnu" and "x86_64-apple-darwin" to ""arm-none-eabi-gcc" tools
This is intended to fix [this issue with the Apollo3 Arduino core](https://github.com/sparkfun/Arduino_Apollo3/issues/4) and also anticipate a similar issue with Mac OSX.

However, this is a shot in the dark -- no new binaries are added and instead the new host specifications just link to the same files. Hoping that these binaries can run across the different architectures. If not then specific binaries for these hosts may need to be added.