Define custom cache menu title for SAMD51 boards
The "SparkFun SAMD51 MicroMod" and "SparkFun SAMD51 Thing Plus" boards have a custom "cache" board option that controls
flash caching:
https://learn.adafruit.com/adafruit-grand-central/adapting-sketches-to-m0#cache-3026398-25

From the Arduino platform specification:
https://arduino.github.io/arduino-cli/latest/platform-specification/#custom-board-options

> We must first define a set of menu.MENU_ID=Text properties. Text is what is displayed on the GUI for every custom menu
> we are going to create and must be declared at the beginning of the boards.txt file:

This menu.cache property was missing (I guess it was not transferred over from Adafruit SAMD Boards):
https://github.com/adafruit/ArduinoCore-samd/blob/1.6.5/boards.txt#L17

This resulted in the option not being controllable via a Tools > Cache menu in the Arduino IDE, as is normally the case
for custom board options. More significant, it breaks Arduino Pro IDE (which is more strict about
specification-compliance) when one of those boards is selected.